update some features

+++ src/views/Dashboard/Ujian/IndexView.vue
@@ -11,8 +11,8 @@
             </div>
             <div class="row">
                 <div class="col-lg-3">
-                    <div class="card">
-                        <div class="card-body" style="position: relative;">
+                    <!-- <div class="card"> -->
+                        <!-- <div class="card-body" style="position: relative;">
                             <div class="d-flex align-items-center justify-content-between">
                                 <div>
                                     <h3>4 Ujian</h3>
@@ -25,8 +25,8 @@
                                 <div class="expand-trigger">
                                 </div>
                             </div>
-                        </div>
-                    </div>
+                        </div> -->
+                    <!-- </div> -->
                 </div>
                 <!-- <div class="col-lg-3">
                             <div class="card">                            
@@ -55,7 +55,6 @@
                         </div>
                     </div>
                 </div>
-
             </div>
             <div class="card">
                 <div class="card-body">
@@ -78,13 +77,9 @@
                                     <td>{{ujian.mapel.nama_mapel}}</td>
                                     <td>{{ujian.kelas.nama_kelas}}</td>
                                     <td>
-                                        <router-link :to="{name: 'ujian.detail',params:{id:ujian.id_ujian}}"
-                                            class="btn btn-info btn-sm">Detail</router-link>
-                                        <button class="btn btn-danger btn-sm"
-                                            @click.prevent="destroy(ujian.id_ujian, index)">Delete</button>
-                                        <router-link :to="{name: 'soal.index', params:{id:ujian.id_ujian}}"
-                                            class="btn btn-warning btn-sm"><i class="icon-settings feather"></i>Atur
-                                            Soal</router-link>
+                                        <router-link :to="{name: 'ujian.detail',params:{id:ujian.id_ujian}}" class="btn btn-info btn-sm">Detail</router-link>
+                                        <button class="btn btn-danger btn-sm" @click.prevent="destroy(ujian.id_ujian, index)">Delete</button>
+                                        <router-link :to="{name: 'soal.index', params:{id:ujian.id_ujian}}" class="btn btn-warning btn-sm"><i class="icon-settings feather"></i>Atur Soal</router-link>
                                     </td>
                                 </tr>
                             </tbody>
@@ -94,11 +89,11 @@
             </div>
         </div>
         <!-- Footer START -->
-        <div class="footer fixed-bottom ">
+        <!-- <div class="footer fixed-bottom ">
             <div class="footer-content justify-content-md-end">
                 <p class="mb-0">Copyright © 2022. All rights reserved.</p>
             </div>
-        </div>
+        </div> -->
         <!-- Footer End -->
     </div>
 </template>
@@ -107,32 +102,62 @@
 
 import axios from 'axios'
 import {onMounted, ref} from 'vue'
+import Swal from 'sweetalert2'
 
 export default {
-    // 
+
     setup(){
         let ujian = ref([]);
 
         onMounted(() => {
             //get ujian from api
+               console.log("ujian id : " + localStorage.ujian_id)
             axios.get('/api/ujian')
             .then((result) => {
                 ujian.value = result.data
+
             }).catch((err) => {
                 console.log(err.response)
             });
         });
 
-        function destroy(id,index){
-            axios.delete(
-                `/api/ujian/${id}`
-            )
-            .then(() => {
-                ujian.value.data.splice(index,1)
-                window.location.reload();
-            }).catch((err) => {
-                console.log(err.response.data);
-            });
+        async function destroy(id,index){
+            Swal.fire({
+                title: 'Hapus Data Ujian',
+                text: "Yakin dihapus ?",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Hapus !'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    axios.delete(
+                        `/api/ujian/${id}`
+                    )
+                        .then(() => {
+                            ujian.value.data.splice(index, 1)
+                            Swal.fire({
+                                title: 'Berhasil!',
+                                text: 'Ujian Berhasil dihapus',
+                                icon: 'success',
+                                confirmButtonText: 'Lanjut !'
+                                 }
+                            )
+                            window.location.reload();
+                        }).catch((err) => {
+                            console.log(err.response.data);
+                            Swal.fire({
+                                title: 'Gagal!',
+                                text: 'Kosongkan Soal terlebih dahulu !',
+                                icon: 'error',
+                                confirmButtonText: 'Lanjut !'
+                            }
+                            )
+                        });
+                    
+                }
+            })
         } 
 
         return {
