+++ src/views/Guru/Ujian/IndexView.vue
@@ -1,17 +1,17 @@
 <template>
     <div class="content">
-                <div class="main">                    
-                    <div class="page-header">
-                        <h4 class="page-title">Data Ujian</h4>
-                        <div class="breadcrumb">
-                           <span class="me-1 text-gray"><i class="feather icon-home"></i></span>
-                           <div class="breadcrumb-item">Home</div>
-                           <div class="breadcrumb-item">Ujian</div>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="col-lg-3">
-                            <div class="card">                            
+        <div class="main">
+            <div class="page-header">
+                <h4 class="page-title">Data Ujian</h4>
+                <div class="breadcrumb">
+                    <span class="me-1 text-gray"><i class="feather icon-home"></i></span>
+                    <div class="breadcrumb-item">Home</div>
+                    <div class="breadcrumb-item">Ujian</div>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-lg-3">
+                    <!-- <div class="card">                            
                                 <div class="card-body" style="position: relative;">
                                     <div class="d-flex align-items-center justify-content-between">
                                         <div>
@@ -25,9 +25,9 @@
                                         </div>
                                     </div>
                                 </div>
-                            </div>
-                        </div>
-                        <!-- <div class="col-lg-3">
+                            </div> -->
+                </div>
+                <!-- <div class="col-lg-3">
                             <div class="card">                            
                                 <div class="card-body" style="position: relative;">
                                     <div class="d-flex align-items-center justify-content-between">
@@ -44,63 +44,74 @@
                                 </div>
                             </div>
                         </div> -->
-                        <div class="col-lg-9">
-                            <div class="">                            
-                                <div class="card-body" style="position: relative;">
-                                     <div class="d-grid gap-2 d-md-flex justify-content-md-end">
-                                        <router-link :to="{name: 'guru.ujian.create'}" class="btn btn-primary me-md-end" type="button"><i class="icon-plus-circle feather"></i>Tambah</router-link>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        
-                    </div>
-                    <div class="card">
-                        <div class="card-body">
-                            <h4>Data Ujian</h4>
-                            <div class="mt-4">
-                                <table id="data-table" class="table data-table">
-                                    <thead>
-                                        <tr>
-                                            <th>No.</th>
-                                            <th>Nama Ujian</th>                                           
-                                            <th>Mata Pelajaran</th>                                           
-                                            <th>Kelas</th>                                           
-                                            <th>Action</th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        <tr v-for="(ujian, index) in ujian.data" :key="index">
-                                            <td>{{++index}}</td>
-                                            <td>{{ujian.nama_ujian}}</td>
-                                            <td>{{ujian.mapel.nama_mapel}}</td>
-                                            <td>{{ujian.kelas.nama_kelas}}</td>
-                                            <td>
-                                                <router-link :to="{name: 'guru.ujian.detail',params:{id:ujian.id_ujian}}" class="btn btn-info btn-sm">Detail</router-link>
-                                                <button class="btn btn-danger btn-sm" @click.prevent="destroy(ujian.id_ujian, index)">Delete</button>
-                                                <router-link :to="{name: 'guru.soal.index', params:{id:ujian.id_ujian}}" class="btn btn-warning btn-sm"><i class="icon-settings feather"></i>Atur Soal</router-link>
-                                            </td>
-                                        </tr>                                                                                                 
-                                    </tbody>                                   
-                                </table>
+                <div class="col-lg-9">
+                    <div class="">
+                        <div class="card-body" style="position: relative;">
+                            <div class="d-grid gap-2 d-md-flex justify-content-md-end">
+                                <router-link :to="{name: 'guru.ujian.create'}" class="btn btn-primary me-md-end"
+                                    type="button"><i class="icon-plus-circle feather"></i>Tambah</router-link>
                             </div>
                         </div>
                     </div>
                 </div>
-                <!-- Footer START -->
-                <div class="footer fixed-bottom ">
-                    <div class="footer-content justify-content-md-end">
-                        <p class="mb-0">Copyright © 2022. All rights reserved.</p>
-                    </div>
-                </div>
-                <!-- Footer End -->
+
             </div>
+            <div class="card">
+                <div class="card-body">
+                    <h4>Data Ujian</h4>
+                    <div class="mt-4">
+                        <table id="data-table" class="table data-table">
+                            <thead>
+                                <tr>
+                                    <th>No.</th>
+                                    <th>Nama Ujian</th>
+                                    <th>Mata Pelajaran</th>
+                                    <th>Kelas</th>
+                                    <th>Action</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                <tr v-for="(ujian, index) in ujian.data" :key="index">
+                                    <td>{{++index}}</td>
+                                    <td>{{ujian.nama_ujian}}</td>
+                                    <td>{{ujian.mapel.nama_mapel}}</td>
+                                    <td>{{ujian.kelas.nama_kelas}}</td>
+                                    <td>
+                                        <router-link :to="{name: 'guru.ujian.detail',params:{id:ujian.id_ujian}}"
+                                            class="btn btn-info btn-sm"><i class="icon-clipboard feather"></i>Detail
+                                        </router-link><br>
+                                        <button class="btn btn-danger btn-sm"
+                                            @click.prevent="destroy(ujian.id_ujian, index)"><i
+                                                class="icon-trash-2 feather"></i>Delete</button><br>
+                                        <router-link :to="{name: 'guru.soal.index', params:{id:ujian.id_ujian}}"
+                                            class="btn btn-warning btn-sm"><i class="icon-settings feather"></i>Atur
+                                            Soal</router-link><br>
+                                        <router-link :to="{name: 'guru.nilai.index', params:{id:ujian.id_ujian}}"
+                                            class="btn btn-success btn-sm"><i class="icon-list feather"></i>Nilai
+                                        </router-link><br>
+                                    </td>
+                                </tr>
+                            </tbody>
+                        </table>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <!-- Footer START -->
+        <!-- <div class="footer fixed-bottom ">
+            <div class="footer-content justify-content-md-end">
+                <p class="mb-0">Copyright © 2022. All rights reserved.</p>
+            </div>
+        </div> -->
+        <!-- Footer End -->
+    </div>
 </template>
 
 <script>
 
 import axios from 'axios'
 import {onMounted, ref} from 'vue'
+import Swal from 'sweetalert2'
 
 export default {
     // 
@@ -123,16 +134,43 @@ export default {
             });
         });
 
-        function destroy(id,index){
-            axios.delete(
-                `/api/ujian/${id}`
-            )
-            .then(() => {
-                ujian.value.data.splice(index,1)
-                window.location.reload();
-            }).catch((err) => {
-                console.log(err.response.data);
-            });
+        async function destroy(id, index) {
+            Swal.fire({
+                title: 'Hapus Data Ujian',
+                text: "Yakin dihapus ?",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Hapus !'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    axios.delete(
+                        `/api/ujian/${id}`
+                    )
+                        .then(() => {
+                            ujian.value.data.splice(index, 1)
+                            Swal.fire({
+                                title: 'Berhasil!',
+                                text: 'Ujian Berhasil dihapus',
+                                icon: 'success',
+                                confirmButtonText: 'Lanjut !'
+                            }
+                            )
+                            window.location.reload();
+                        }).catch((err) => {
+                            console.log(err.response.data);
+                            Swal.fire({
+                                title: 'Gagal!',
+                                text: 'Kosongkan Soal terlebih dahulu !',
+                                icon: 'error',
+                                confirmButtonText: 'Lanjut !'
+                            }
+                            )
+                        });
+
+                }
+            })
         } 
 
         return {
